day-11 codes uploaded

+++ case-study/case_study_queries/sp_DeleteEmployeeRecord.sql
@@ -1,6 +1,17 @@
 ﻿create procedure DeleteEmployeeRecord
-@id int
+@id int,
+@res int out
 as
 begin
-delete from employees where employeeid=@id
+	declare @count int;
+	SELECT  @count = count(*) FROM    employees  WHERE   employeeid = @id
+	if @count=1
+		begin
+			delete from employees where employeeid =@id
+			set @res = 1
+		end
+	else
+		begin
+			set @res = 0
+		end
 end
